AP-9889: Shorten string representation already on Python side
The string representation of a pickled object is used in the view and summary of the port object (spec).
For some objects, the string representation is quite large as it includes the objects' entire contents. (Gets even larger than the actual pickled byte representation.)
In order to reduce the (de)serialization load (both between Python and Java, and to/from disk) and since we always truncate the string to a manageable length (1000 characters) for display, we now shorten the string already on Python side.