AP-10584: Fix writing tables with zero rows (Arrow)
A table of row count zero is interpreted as a table of length zero whose columns only contain missing values.
Until now, a dummy buffer needed to be sent to Java for such columns due to limitations on Apache Arrow side, which caused problems in the case of length zero (as the dummy effectively implies length one). This dummy seems not to be required any more. Therefore we simply remove the corresponding code.