Fix registration creating multiple instances of singleton
There was a flaw in our implementation of the singleton pattern:
Registering extensions from within the constructor of the singleton
class may trigger the activation of other bundles that may try to
access the registry's singleton instance before it is fully constructed,
therefore creating multiple instances of the registry.
This is fixed by separating instance construction from registration.