AP-13724: Support setting options after starting kernel
This prepares the kernel for the introduction of a pre-launch mechanism
based on queues to speed up Python node execution.

To make a pre-launched kernel as universally usable as possible, most of
its configuration needs to be deferred until it is actually used. We
therefore separate PythonCommand (the only essential information a
kernel needs to launch) from PythonKernelOptions and support setting the
rest of the options (serializers, etc.) at a later point in time.

AP-13724 (Proactively launch Python to speed up future executions)