AP-13724: Use cached test results for subsets of modules
So far, cached test results of previous test runs were only used if the
sets of additional modules of the old run and the new run were equal.
Now, they are also reused if the set of modules of the new run is fully
contained inside the one of the old run.

While this change makes sense in general, it was motivated by the new
possibility of starting a "bare-bone" Python kernel (without a
serializer, etc. set) and configuring it afterwards. This results in
installation tests at two different sites: once when constructing the
kernel (testing no additional modules) and when setting the options
(testing additional modules). The first installation tests during the
lifetime of the Python plugin are initiated by its Activator and run for
the two environments configured in the Preferences. These tests include
the additional modules required by the configured serializer. By
introducing this new "subset matching" in the installation tester, the
results of these first tests can also be reused by the first "bare-bone"
Python kernel even though it does not specify the modules of the
serializer.

AP-13724 (Proactively launch Python to speed up future executions)