AP-13724: Introduce Python kernel queue
Holds a number of pre-launched Python kernels. Allows clients to obtain
these kernels, reducing the startup-overhead that is involved when
employing Python in KNIME nodes to zero, in the best case.
The underlying Apache Commons Pool implementation handles most of the
life-cycle management of the enqueued kernels and also reacts to
different work loads. This comprises dynamically adding kernels to the
queue when it is exhausted as well as closing and evicting kernels when
the maximum capacity of the pool is reached or kernels have expired.

AP-13724 (Proactively launch Python to speed up future executions)