AP-13917: Expose default parameters of kernel queue
They are going to be used as defaults in the preferences.

Also change the resolution of the expiration parameter to minutes since
we only evict expired entries once a minute anyway. We should not give
the user the impression that changing this parameter in the sub-minutes
range would have any discernible effect.

AP-13917 (Make Python kernel queue parameters configurable)