AP-13917: Make kernel queue (re)configurable
Reconfiguring the queue is simply done by closing the current singleton
instance and replacing it by one that was constructed according to the
new configuration. This requires some synchronization on the class
level but is in general easier than mutating the queue while it is being
in use.

AP-13917 (Make Python kernel queue parameters configurable)