DEVOPS-418: Rework failing serialization tests
Mostly change the tested data type from float32 (which KNIME does not
support natively anyway) to int32. The float32-based tests have never
passed in the first place for conceptual reasons (not bugs), so doing
this is not critical and actually increases test coverage.

Also, it turned out that Flatbuffers and Arrow serialization cannot be
tested in an "offline" fashion (i.e., without reaching out to Python),
since the data formats of the "Java-to-Python" and "Python-to-Java"
directions differ from one another. We therefore turn the abstract base
class SerializationTest into a utility to allow the different test
implementations to choose which modes (offline vs online) they are
testing.

DEVOPS-418 (Configure Python environments and tests in new build system)