AP-14610: Use setResizeWeight only
Previously, setDividerLocation(int) was called on the split panes, causing
layout issues on high-dpi displays on Windows. The resize weights need
to be set in any case. In fact, only setting the resize weights suffices
to determine the layout if the preferred sizes of the contained
components are set to zero.

(AP-14610) Python dialogs: Layout seems broken on Windows 10