AP-15256: Null check for String flow variables
Flow variables of type String can carry null values. During
serialisation for python nodes, such a null value would cause an error.
As per this commit, in case of a null value, the value is handled as if
it was a missing value. In effect, this will cause it to be interpreted
as `None` on the python side.

AP-15256 (Empty String Flow Variable crashes Python Serialization)