AP-16015: Fix bugs labeled "scary" by SpotBugs
We were using the string representation of the arrays' identities rather
than their contents.
This is on a code path that is likely neither tested nor used in
production (it is only a fallback for extension types without a
registered Python deserializer). Still worth fixing.

AP-16015 (Core Team Code Cleanup Marcel)