AP-17289: Rework PythonModuleKnime* and PythonPath utils
In an earlier commit, we simplified the structure of our Python source
code directories. Because of this, we do not need to be able to resolve
the paths of individual Python module files anymore but just the path
of each plug-in's base Python source directory.

AP-17289 (Implement putDataTable for real)