Prevent Python files from being copied to bin/ or target/ during build
As the duplicates can be really confusing when searching for or editing
modules in e.g. an IDE.

This is done by standardizing the project, class-path, and build
settings of all python3 projects once more. In particular, make sure
that:

- There is no PyDev build command registered in .project
- There is no class-path entry and no source entry in build.properties
  pointing to the Python source folder
- The Python source folder is (still) included in the binary in
  build.properties

This seems to fix the problem at least on my machine.