AP-17624: Add logical type version to metadata
The factory versions are hard coded in Python and need to match those on
the Java side. We have added versioning to the ExtensionArrowColumnDataFactory
on the Java side that maps to the LogicalTypeExtensionType version in
Python and hence needs to be added to the version string.

Added version fields to the ExtensionTypes to make the connection more
obvious.

AP-17624 (Bump ArrowColumnarDataFactory versions)