AP-17624: Fill in LogicalType for known types when coming from Pandas
We should write all information into the Arrow file from Python and not
just add the Traits later, because we want to be able to go from one
Python node to the next and have all information available.

Currently that is implemented by a simple map of the data types needed
for our tests. But a ToDo is in place to use the DefaultValueFactories
for this mapping.

AP-17624 (Bump ArrowColumnarDataFactory versions)