AP-17055: Add selection panel for kernel back end
And make it part of the Options panel of the dialog.

Also merge in the panel's base class. The Options panel is its only
implementation and the code separation was more annoying than helpful.

Also fix some Sonar issues and compiler warnings.

AP-17055 (UI to switch between "old" and "new" PythonKernel and API)