AP-18069: Filter flow variables for Python also when populating flow variable pane
Otherwise the dialog would display all flow variables and allows
inserting them into the code via double click, however the variable is
not really available in Python. Now we are more consistent with the
behavior of the Python2 scripting node in that we only show those
variables that are actually available.

AP-18069 (Python Script (Labs) node swallows conda environment variable)