AP-18067: Convert object-type columns of pd.DataFrames with only missing values to void
When we retrieve columns with only missing values from pandas, we apply
the VoidValueFactory logical type to make it work in KNIME. Arrow has a
bug when it comes to wrapping null arrays in ExtensionTypes, so we
circumvent that by creating a null vector with the type we want.

For lists full of missing values, we now return ListValueFactory with
VoidValueFactory elements, which KNIME understands properly.
The creation of the appropriate Arrow array is a little more involved
and thus covered by a separate test case.

+ Added more test cases for primitive type mappings.

AP-18067 (Python Table with Void Column (all values are None) throws
error)