AP-18067: Fix unwrapping of unspecific ListValueFactory type, handle List<Void>
Unwrapping needs to handle the case that the storage of an extension
type is a list, and if that list contains extension types, these need to
be unwrapped as well.

Showed up now for the first time because the previous commits enables
saving List of primitive-logical-types (List<void>).

AP-18067 (Python table with Void column (all values are None) throws
Error)