AP-17993: Add checking for being in a single-line string for old Jedi API
For Python environments with Jedi version before 0.16.0, autocompletion falls back to the old Jedi API. This means that it does not automatically check if the cursor is within a single- or multi-line string. This check is now done manually for single-line strings by counting quotation marks in the current line.

AP-17993 (Python auto-completion opens in comments)