AP-18042: Improve error msg when unsupported types are returned from Python
There are two places in the code where such an error can occur, once
when determining the schema, and once when putting data into the table.

All data that gets written goes through wrap_primitive_types() and for
KNIME we need that each data type is a logical type. So we check that
all data types returned to Python have a logical type or give a
comprehensive error message.

AP-18042 (Python: knio.write_table throws error, if cell type is not supported by PyArrow)