AP-17775: Add Python table schema
To implement configure() of a Python node, we add a Python equivalent to the DataTableSpec.
This includes a type system, for which we took a lot of inspiration from pyarrow.

This schema can be read and written to JSON to communicate with KNIME, and we have a unittest
including a JSON file serialized from KNIME that we can read in Python.

AP-17775 (Deep Dive: Node API)