AP-17775: Add test to experiment with functional Python Node API
This includes a Table that combines ReadTable and WriteTable from knime_table.
A common interface (Columnar) is added for Schema and Table, which provides
all operations that we need for a functional API. This way the same code
can be used for configure (with Schemas) and execute (with Tables), making
it really easy to write functional nodes.

AP-17775 (Deep Dive: Node API)