AP-18690: Add Pandas column conversion extension mechanism
For some special extensions (geospatial) we might want to perform some
special treatment of a column after converting it to a Pandas DataFrame
or before converting it into a PyArrow table. To allow this, we have
introduced the ToPandas- and FromPandasColumnConverter interfaces
and means to register those.

Even though they are specific to Pandas, these converter interfaces are
currently located in knime_types because that is also where
PythonValueFactories are registered from the outside.

We did not add a Java/OSGi ExtensionPoint for those converters yet as
we do not expect that this is a feature that is needed often, and we
can already register them from within Python modules.

AP-18690 (GeoSpatial Python Extension Type)