AP-18690: Add Python GeoSpatial type test
We add a testcase for the GeoSpatial types here because the test needs
access to some internals (moc an ArrowDataSource...) and also serves as
test for the column conversion extension mechanism.

However, it only works if knime-geospatial is checked out right next to
knime-python. We could consider moving this test over to knime-geospatial.

AP-18690 (GeoSpatial Python Extension Type)