AP-18103 Python: knio.BatchWriteTable.append cannot process lists and sets
Switched the return statement for non-empty dataframes in kap.pandas_df_to_arrow from returning a RecordBatch or pa.Table to only a pa.Table. This was because our ExtensionArrays are sometimes backed by ChunkedArrays with ExtensionTypes which are not supported by RecordBatches.

AP-18103 (Python: knio.BatchWriteTable.append cannot process lists and sets)