AP-14756: Add preference page for Python3 scripting node and provide bundling option
We decided to add a new preference page for the Python3 (Labs) Scripting node because
we can strip away a lot of settings which only apply to org.knime.python2 and because
we want to add a new option: the bundled Python environment.

The panel for the bundled Python environment only contains a description because there
is nothing to configure. All other elements are derived and adjusted from the Python2
preference page.

AP-14756 (Make bundled Python an option in the executable tab)