AP-14756: Initialize Python (Labs) preference page from current Python pref page settings
As we have just added this new preference page but had the nodes using values from the
(old) Python pref page before, we initialize the new pref page with values from the old
pref page so that for users with valid configuration nothing changes.

In case Conda was not properly configured but a bundled env is available, we switch to
bundled automatically to make things work out of the box.

If bundled was selected but is not available, we switch back to Conda automatically.

To set the Python (Labs) defaults to values of the Python pref page, we added a PreferencesInitalizer
that loads the PythonPreferences and writes them into the Python (Labs) default preferences.

AP-14756 (Make bundled Python an option in the executable tab)