AP-14756: Use bundled also for first Python execution if available and conda not configured
If a Python Script (Labs) node gets executed and the Python (Labs) preference page has
not been visited before, then it will load the defaults. Those defaults are taken from
the Python preference page since the last commit, but these can never be "Bundled".
Now, if conda is not configured but a bundled env is available, we already set "Bundled"
as default before the preference page is opened up, so that Python Script (Labs) nodes
can be executed out of the box.

AP-14756 (Make bundled Python an option in the executable tab)