AP-18642: Implement Python Table for KNIME node execution
The new Python Table shares a lot of the interface with a table schema,
in fact they can be used as drop in replacement because they support all
the same operations on column level.

The backend implementation is separated, to hide the Arrow logic from our users.

Added a lot of tests for the new Table.