AP-18412: Implement DenseBitVectorValueFactory in Python
As a first attempt we are using a hex string serialization
and try to use the bitarray package to provide a nice Pythonic
way of accessing the data. This will be improved in AP-19224
because hex strings get truncated on the KNIME side as they are
only meant for displaying apparently.

AP-18412 (Implement RDKit Extension Types)