AP-18963: Display MultiColumnParameter in dialog even if not configured
So far the developer had to set something in configure in order for a
MultiColumnParameter to be displayed. Now we set the include list to []
if the value is None during dialog generation. We keep None as the
indication for the unconfigured state, so the API is stable and existing
nodes don't have to be adjusted.

AP-18963 (Column Parameters are not displayed)