AP-19147: Allow configure to return None
Both for individual specs as well as in general (which is equivalent to
not returning at all). If None is returned for any port then all nodes
downstream of this node can't be configured unless configure returns a
spec during a later configure call or after execute.

AP-19147 (Python Node Extension: None is not an allowed output of
configure)