AP-19139: Handle parameter storage depending on parameter_group type
The two types being parameter_groups used as descriptor and
parameter_groups used via composition. If used as a descriptor, the host
object stores the parameters. If used via composition, the
parameter_group itself has to store the parameters because the host
object is not available when a parameter is set via Python. This implies
that parameter_group instances can't be shared when used via composition
but that shouldn't be a problem since this is in-line with the general
intuition of object oriented programming.

AP-19139 (Python Node Extension: Dialog doesn't open when a new setting
is added)