AP-18412: Add a monkeypatch for the as_py method of Extension Scalars in pyarrow==8
by @jonas.klotz.

PyArrow 8 changed something in their handling of extension arrays and scalars
such that when accessing items of a chunked array, we no longer receive a KnimeExtensionScalar
but a pyarrow ExtensionScalar without a way to control how as_py interprets the underlying data.
This will be changed in PyArrow 9 (we need to add as_py_scalar() to our extension type),
but for the time being we warn if Pyarrow 8 is used or monkey patch their ExtensionScalar
if "forbiddenfruit" is installed. We need this library for monkey patching because ExtensionScalar
is a Cython type which cannot easily be modified in plain Python.

AP-18412 (Implement RDKit Extension Types)