AP-19074: Make sure that input ports are copied into the node when created via node_factory
We already made sure that we inject the ports (collected from decorators and class variables)
into the created instance when using the MyNode constructor manually, but when the node
was created from the KnimeNodeBackend we were missing such a consolidation step. Now both
means of creating node instances behave the same way.

AP-19074 (Cannot parse Python node without any inputs or outputs if not extending PythonNode)