AP-19138: Make error on duplicate rows nicer
A DuplicateKeyException was wrapped as an IOException by using the
IOException(Throwable cause) constructor. This led to an error message
containing the class name of the DuplicateKeyException which looks
strange to the user.
We now use the message of the DuplicateKeyException directly and add
information to the message in the SinkManager.

AP-19138 (Python Node Extension: Uninformative error on duplicate RowKeys)