AP-19215: GeoDataFrame needs to be converted to Pandas DataFrame before returning
Convert Geodataframes to pandas Dataframes in pandas_df_to_arrow, such that the CRS is not lost. Copy the dataframe before converting the columns, such that the conversion process does not reflect in the original dataframe.
Implement 2 Testcases, one to check equality after conversion to table, one to check  that the content is equal and the crs is not lost.

AP-19215: (GeoDataFrame needs to be converted to Pandas DataFrame before returning)