AP-19364: Implement __arrow_ext_scalar_class__ also for dictionary encoded types
For StructDictEncodedLogicalType this means delegating the access to the value,
but for StructDictEncodedType we actually cannot implement this properly because
we do not have access to the full array and cannot look up the dictionary values.

We are in contact with the Arrow / Voltron people to discuss whether/how our
dictionary use case could be supported.

AP-19364 (Use new ExtensionScalar features from PyArrow 9.0)