AP-18612: Add type hints in knime_parameters.py
Added type hints for `version` parameters of functions and class methods.

Clarified that we must always use the installed extension version when calling the `getDialogRepresentation` method in order to obtain the dialogue representation that contains the schema and ui schema that match the set of parameters available in the installed extension version.

AP-18612 (Support versioning of parameters)