AP-19334: Capture stdout/err in Python process
and send it to Java via py4j instead of streams. This way the outputs to
both, stdout and stderr are sent to Java in the correct order and the
streams are not interleaved at the wrong locations.

AP-19334 (Remember stdout/stderr from the last execution and provide it to the frontend)