AP-19372: Only reuse explicitly set Python Conda environments, otherwise use bundled
We were checking whether conda is configured to select "bundled" or "conda" on the
Python (Labs) preference page, however on Windows conda is pretty much always installed
at the default location and therefore our check returns 'Conda is configured' and
the bundled environment did not get selected. However, we only want to select
conda by default if the user had explicitly configured a Python environment to use.
So now we get the Python env configuration from the _old_ Python preference page,
check whether a Python 3 environment has been explicitly selected by the user,
and if that's not the case, then we use the Bundled environment.

AP-19372 (Default for new Windows Installation is not Bundled but Conda