AP-19387: Create a simple infrastructure for determining backward and forward compatibility cases
We compare the version of the extension that the settings were saved with with the version of the installed extension, and determine whether we have a case of backward compatibility or forward compatibility. The user is then notified with a warning of it's the former, with a list of newly added parameters, or with an error message if it's the latter.

AP-19387 (Display warning and error messages when loading Python nodes with a different extension version)