AP-18712: Improved Dictionary Encoding speed
As pyarrow re-instantiates our DictEncodedExtensionArray for every access, thus, recalculating the dictionary keys
everytime, the access was inefficient. A global LRU Cache dictionary is implemented, mapping the memory addresses
of the data contained in each array to its keys.

AP-18712: (Dictionary encoded fields of structs are not unwrapped in Python)