AP-19387: Extract determineCompatibility into a separate interface and add unittest
The determineCompatibility method is now declared in a new VersionedProxy interface, which the NodeConfigurationProxy and PythonNodeProxy extend, and which CloseablePythonNodeProxy implements.

Added a unittest for the `determine_compatibility` Python method that covers forward amd backward compatibility cases with various missing parameters.

AP-19387 (Display warning and error messages when loading Python nodes with a different extension version)