AP-19224: Allow casting Pandas series between extension types
Because Arrow serializes and deserializes the extension type information of
arrays all the time and reconstructs the type from this info, we had to make
sure that a "changed" converter inside a LogicalTypeExtensionType is also remembered
through this serialization roundtrip. We achieved this be introducing a new
ProxyExtensionType which holds the extra information. This extension type
is replaced by the underlying LogicalTypeExtensionType before being saved to disk
and hence before it is read by Java in KNIME.

AP-19224 (Implement Dense Bit and Byte Vector Types)