AP-19556: Allow node extension to use another bundled environment
This is needed for node extensions that use for example the default
bundled python scripting environment.
For now, this feature is undocumented and cannot be used by usual Python
node extensions because the bundling does not support it. However, we
can use it for the testing node extension. Later, we can add support to
the bundling to make it available for other Python node extensions.

AP-19556 (Create a testing Python node extension)