AP-18368: Extract port logic into PortTypeRegistry
This was necessary to check that port types are registered.
Otherwise extensions could suddenly fail once we can have
dedicated port types for the different port types registered
in Python.
Also required some refactoring of the unit tests because the
methods for converting port objects and their specs back and
forth between java and python moved into the PortTypeRegistry
and were used in those tests.