AP-18654: Deprecate knime_io, add knime.scripting.io and use the same table as in Python nodes
We fuse the API so that we use the same API for Tables in both, the scripting node
and pure-Python KNIME nodes.

For that, the table interface and implementation have moved to knime.api.table and
knime._arrow._table respectively. This matches the new file structure we want to have
for the new API anyways.

AP-18654 (Integrate the new Table API also in the Python Script (Labs) node)