Force pref init when Python3ScriptingPreferences are loaded
The preference initializer is called lazily the first time the
preference node is accessed. If there are multiple KNIME nodes in a workflow
that are executed at the same time this can happen in parallel.
However, the loading of the preference node is not thread-safe and
therefore, the node might not have finished loading before being accessed by
one of the threads.