AP-19695: Add feature dependency to columnar backend
The columnar backend is required for the python3.scripting feature and
for the python3.nodes feature because it uses columnar tables for
transferring data. If we do not add the feature dependency the
platform-specific native libraries for lz4 will be missing from the
installation and the columnar backend will not be usable.

AP-19695 (WF created with columnar table backend cannot be saved in KNIME without columnar table backend)