AP-19793: Create markdown parser once per backend
If the "markdown" package is not installed we only complain once per
Python process and not every time we are doing anything. The markdown
parser is not created in the __init__ method of the _KnimeNodeBackend
because the logging is not yet set up when the __init__ method is
executed.

AP-19793 (Log is spammed with warning about markdown not being installed)