AP-20048: Implement column filtering logic in Python, hand ColumnFilterConfig object to user
The full column filter has multiple modes how the user can select columns. As the list of "selected"
columns that the user has seen in the dialog might not be valid at the next configure/execution time
because the spec of the input table changed, we need to be able to apply all the filtering logic
in Python.

Regex, Wildcard and Manual filtering work completely, type filtering currently only works during
configure because we do not have the preferred_value_type available in the tables read from Arrow.

AP-20048 (Support full column filter in Python extensions)