AP-20048: Inject preferred_value_type into table schema during execute
During execute, the schema is read from the arrow table and not coming
from Java, but the preferred_value_type information is not part of the
Arrow table. The table contains information about the types, however,
so we can query the Java side to give us the preferred_value_type and
inject that into the column metadata.

This is needed so that the column filter's type filtering mode works
during execute.

AP-20048 (Support full column filter in Python extensions)