UIEXT-1261: Default dialog settings wrapped in extra 'data' property
I.e. extract it when parsing the data for the python-java backend when
received from the frontend.
But ignore the 'data'-property when the settings are received from
python.

Also a workaround employed to avoid NodeSettings.copyTo where a
'NodeAndVariableSettings'-proxy is the target (because that doesn't
work since it can't be cast to ConfigBase).

UIEXT-1261 (Communicate selected flow variables on \)