AP-21485: Differentiate state between running all and selected
We locally remembered which button was clicked in PythonEditorControls.
This is used to show the cancel text on the correct button. However, an
execution can also be triggered from the view placeholder. In this case,
the local state was not changed which caused the cancel text to be
shown on the wrong button (potentially).