AP-22103: Reset m_view also if m_consoleOutputStorage is null
If the node was saved with 5.1 m_consoleOutputStorage is null. However,
the view is still present for the "Python View" node. If we do not set
the view to `Optional.empty()` it will be deleted from the original
workflow with the next node execution.

AP-22103 (Python Views don't open after unsaved re-execute)