AP-22697: Use the live value of the exeutable selection for saving the settings
Before, we would only update the value in the store if the settings page
was unmounted. This is not good enough anymore, because users can now apply
the dialog also without unmounting the preference page.

AP-22697 (Python environment settings not applied if "Ok" pressed while inside the "Set Python environment" page)