AP-22425: Init dialog NotInWorkflowWriteFileStoreHandler with workflow data repo
Using `NotInWorkflowWriteFileStoreHandler.create` will cause incoming file
stores to be moved to the `NotInWorkflowFileStoreHandler` when copying the
row-based table to a columnar table. Later when closing the dialog this hander
is cleared and therefore the file store is lost.

AP-22425 (Python node dialog moves and resets file stores of incoming row-based table)