AP-24827: Delete the timzone from the value for saving it
The timezone is constant (given by the parameter). The values are saved
and presented to the user as date&time values without timezone
information (because the user cannot change the timezone).

AP-24827 (Python nodes: DateTimeParameter does not show configured value in dialog)
AP-24836 (Using DateTimeParameter with timezone causes time-shifts related to the system timezone)