AP-25011: Fix regression of knime-core-ui when selecting 'None' column.
Updated parameter.py for single column selection since knime-core-ui accepts '{id: <none>, text: None}' in key 'possibleValues'. This in turn required overriding methods _to_dict() and _from_dict() to support '<none>' values when encountered. Note that the old attribute for setting 'showNoneColumn' has not been touched and will still be part of json_forms_schema.
The inject method handles the case when no column is selected, or the list of columns to select from is empty. This is done separately to avoid it clashing with '<none>' value that is to be used only if the 'None' option is introdued in column.

AP-25011 (Column Parameter's attribute in knime.extension include_none_column no longer works)