Properly skip Geospatial unit tests for Pytest 9.0
We were using a check for the availability of geo types in the setUpClass method,
followed by raising a unittest.SkipTest exception if the types were not found.

This behavior apparently worked with pytest up to version 8, but not in version 9.

Thus we now implement the test skipping by using the @pytest.mark.skipif decorator